% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhatR.R
\name{add.dependency}
\alias{add.dependency}
\title{Private function that adds a package to the list of dependencies
that will be installed on the ScienceOps server}
\usage{
add.dependency(name, src, version)
}
\arguments{
\item{name}{name of the package to be installed}

\item{src}{source that the package is installed from (CRAN or github)}

\item{version}{version of the package}
}
\description{
Private function that adds a package to the list of dependencies
that will be installed on the ScienceOps server
}

