% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhatR.R
\name{add.import}
\alias{add.import}
\title{Private function that adds a package to the list of imports}
\usage{
add.import(name)
}
\arguments{
\item{name}{name of the package to be added}
}
\description{
Private function that adds a package to the list of imports
}

