% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yhatR.R
\name{capture.src}
\alias{capture.src}
\title{Private function for catpuring the source code of model}
\usage{
capture.src(funcs)
}
\arguments{
\item{funcs}{functions to capture, defaults to required yhat model functions}
}
\description{
Private function for catpuring the source code of model
}

