% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhatR.R
\name{check.dependencies}
\alias{check.dependencies}
\title{Checks dependencies and makes sure all are installed.}
\usage{
check.dependencies()
}
\description{
Checks dependencies and makes sure all are installed.
}
