% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhatR.R
\name{check.image.size}
\alias{check.image.size}
\title{Private function for checking the size of the user's image.}
\usage{
check.image.size()
}
\description{
Private function for checking the size of the user's image.
}

