\name{documentColumn}
\alias{documentColumn}
\title{Documents a column from a given data.frame}
\usage{
  documentColumn(df, col)
}
\arguments{
  \item{df}{the data.farme frame you're documenting}

  \item{col}{the column as a string that you're
  documenting}
}
\description{
  Generates documentation for a column.
}

