% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.https}
\alias{is.https}
\title{Private predicate function that checks if the protocol of a url
is https.}
\usage{
is.https(x)
}
\arguments{
\item{x}{is a url string}
}
\description{
Private predicate function that checks if the protocol of a url
is https.
}

