% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhatR.R
\name{yhat.batchDeploy}
\alias{yhat.batchDeploy}
\title{Deploy a batch model to Yhat servers}
\usage{
yhat.batchDeploy(job_name, confirm = TRUE)
}
\arguments{
\item{job_name}{name of batch job}

\item{confirm}{boolean indicating whether to prompt before deploying}
}
\description{
This function takes model.transform and model.predict and creates
a model on Yhat's servers which can be called from any programming language
via Yhat's REST API (see \code{\link{yhat.predict}}).
}
\examples{
yhat.config <- c(
 username = "your username",
 apikey = "your apikey",
 env = "http://sandbox.yhathq.com/"
)
model.predict <- function() {
  name <- "ross"
  greeting <- paste("Hello", name)
  print(greeting)
}
\dontrun{
yhat.batchDeploy("helloworld", confirm=FALSE)
}
}
\keyword{deploy}

