\name{yhat.document}
\alias{yhat.document}
\title{Document a model using Yhat}
\usage{
  yhat.document(model, version, df)
}
\arguments{
  \item{model}{name of your model}

  \item{version}{version of your model}

  \item{df}{example data for your model}
}
\description{
  Takes a specific model along with data that is required
  to execute a prediction and generates a
  webapp/documentation.
}
\keyword{document}

