\name{yhat.login}
\alias{yhat.login}
\title{A function for logging into Yhat's api.}
\usage{
  yhat.login(username, apikey)
}
\arguments{
  \item{username}{Your Yhat username}

  \item{apikey}{Your Yhat apikey}
}
\description{
  A function for logging into Yhat's api.
}
\examples{
yhat.login("hmardukas", "abcd1234")
}

