% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhatR.R
\name{yhat.ls}
\alias{yhat.ls}
\title{Private function for determining model dependencies}
\usage{
yhat.ls(batchMode = FALSE)
}
\arguments{
\item{batchMode}{boolean to capture yhat.batch code for a batch job}
}
\description{
List all object names which are dependencies of `model.transform`
and `model.predict` or `yhat.batch` if this is a batch mode deploy
}

