\name{yhat.spider.func}
\alias{yhat.spider.func}
\title{Private function for gathering global variables required by a function}
\usage{
  yhat.spider.func("my.func")
}
\arguments{
  \item{func.name}{/string name of function you want to spider}
}
\description{
  Private function for gathering global variables required by a function
}

