% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhatR.R
\name{yhat.transform_from_example}
\alias{yhat.transform_from_example}
\title{Generates a model.transform function from an example input data.frame.
Handles columns which need to be type casted further after the initial JSON
to Robject such as factors and ints.}
\usage{
yhat.transform_from_example(df)
}
\arguments{
\item{df}{A data.frame object which mirrors the kind of input to the model.}
}
\description{
Generates a model.transform function from an example input data.frame.
Handles columns which need to be type casted further after the initial JSON
to Robject such as factors and ints.
}
\examples{
\dontrun{
model.transform <- yhat.transform_from_example(iris)
}
}

