% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhatR.R
\name{yhat.unload}
\alias{yhat.unload}
\title{Removes a library from the Yhat model's dependency list}
\usage{
yhat.unload(name)
}
\arguments{
\item{name}{of the package to be removed}
}
\description{
Removes a library from the Yhat model's dependency list
}
\examples{
\dontrun{
yhat.unload("wesanderson")
}
}
