% Generated by roxygen2 (4.0.0): do not edit by hand
\name{capture.src}
\alias{capture.src}
\title{Private function for catpuring the source code of model}
\usage{
capture.src(funcs)
}
\arguments{
\item{funcs}{functions to caputre, defaults to required yhat model functions}
}
\description{
Private function for catpuring the source code of model
}

