% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhatR.R
\name{is.https}
\alias{is.https}
\title{Private predicate function that checks if the protocol of a url
is https.}
\usage{
is.https(x)
}
\arguments{
\item{x}{is a url string}
}
\description{
Private predicate function that checks if the protocol of a url
is https.
}
