% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/yhatR.R
\name{yhat.get}
\alias{yhat.get}
\title{Private function for performing a GET request}
\usage{
yhat.get(endpoint, query = c())
}
\arguments{
\item{endpoint}{/path for REST request}

\item{query}{url parameters for request}
}
\description{
Private function for performing a GET request
}

