% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yhatR.R
\name{yhat.post}
\alias{yhat.post}
\title{Private function for performing a POST request}
\usage{
yhat.post(endpoint, query = c(), data, silent = TRUE)
}
\arguments{
\item{endpoint}{/path for REST request}

\item{query}{url parameters for request}

\item{data}{payload to be converted to raw JSON}

\item{silent}{should output of url to console be silenced?
Default is \code{FALSE}.}
}
\description{
Private function for performing a POST request
}

