% Generated by roxygen2 (4.0.1): do not edit by hand
\name{yhat.post}
\alias{yhat.post}
\title{Private function for performing a POST request}
\usage{
yhat.post(endpoint, query = c(), data, silent = FALSE)
}
\arguments{
\item{endpoint}{/path for REST request}

\item{query}{url parameters for request}

\item{data}{payload to be converted to raw JSON}

\item{silent}{should output of url to console be silenced?
Default is \code{FALSE}.}
}
\description{
Private function for performing a POST request
}

