\name{yhat.post}
\alias{yhat.post}
\title{Private function for performing a POST request}
\usage{
  yhat.post(endpoint, query = c(), data)
}
\arguments{
  \item{endpoint}{/path for REST request}

  \item{query}{url parameters for request}

  \item{data}{payload to be converted to raw JSON}
}
\description{
  Private function for performing a POST request
}

