% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yhatR.R
\name{yhat.predict_raw}
\alias{yhat.predict_raw}
\title{Calls Yhat's REST API and returns a JSON document containing both the prediction
and associated metadata.}
\usage{
yhat.predict_raw(model_name, data, model_owner, raw_input = FALSE,
  silent = TRUE)
}
\arguments{
\item{model_name}{the name of the model you want to call}

\item{data}{input data for the model}

\item{model_owner}{the owner of the model [optional]}

\item{raw_input}{when true, incoming data will NOT be coerced into data.frame}

\item{silent}{should output of url to console (via \code{yhat.post})
be silenced? Default is \code{FALSE}.}
}
\description{
Calls Yhat's REST API and returns a JSON document containing both the prediction
and associated metadata.
}
\examples{
yhat.config <- c(
 username = "your username",
 apikey = "your apikey"
)
\dontrun{
yhat.predict_raw("irisModel", iris)
}
}

