% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/yhatR.R
\name{yhat.scaffolding}
\alias{yhat.scaffolding}
\title{Quick function for setting up a basic scaffolding of functions for deploying on Yhat.}
\usage{
yhat.scaffolding()
}
\description{
Quick function for setting up a basic scaffolding of functions for deploying on Yhat.
}
\examples{
yhat.scaffolding()
}

