\name{yhat.scaffolding}
\alias{yhat.scaffolding}
\title{Quick function for setting up a basic scaffolding of functions for deploying on Yhat.}
\usage{
  yhat.scaffolding()
}
\description{
  Quick function for setting up a basic scaffolding of
  functions for deploying on Yhat.
}
\examples{
yhat.scaffolding()
}

