% Generated by roxygen2 (4.0.1): do not edit by hand
\name{yhat.show_models}
\alias{yhat.show_models}
\title{Shows which models you have deployed on Yhat.}
\usage{
yhat.show_models()
}
\description{
This function queries the Yhat API and finds the models that have been deployed
for your account.
}
\examples{
yhat.config <- c(
 username = "your username",
 apikey = "your apikey",
 env = "http://cloud.yhathq.com/"
)
\dontrun{
yhat.show_models()
}
# some output here
#    username className                  name version
# 1      greg                 MySMSClassifier       1
# 2      greg                 MySMSClassifier       2
# 3      greg                 MySMSClassifier       3
# 4      greg                 MySMSClassifier       4
}

