% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yhatR.R
\name{yhat.spider.block}
\alias{yhat.spider.block}
\title{Private function for recursively looking for variables}
\usage{
yhat.spider.block(block, defined.vars = c())
}
\arguments{
\item{block}{code block to spider}

\item{defined.vars}{variables which have already been defined within the
scope of the block. e.g. function argument}
}
\description{
Private function for recursively looking for variables
}

