% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yhatR.R
\name{yhat.spider.func}
\alias{yhat.spider.func}
\title{Private function for spidering function source code}
\usage{
yhat.spider.func(func.name)
}
\arguments{
\item{func.name}{name of function you want to spider}
}
\description{
Private function for spidering function source code
}

